Fix compability issues

+++ src/ui/src/infoline.cpp
@@ -106,7 +106,7 @@ void InfoLine::contextMenuEvent( QContextMenuEvent* event )
              [this]( auto ) { QApplication::clipboard()->setText( this->selectedText() ); } );
 
     connect( selectAll, &QAction::triggered,
-             [this]( auto ) { setSelection( 0, this->text().length() ); } );
+             [this]( auto ) { setSelection( 0, static_cast<int>(this->text().length() )); } );
 
     menu.exec( event->globalPos() );
 }
